Final Code

+++ src/main/java/com/jilds/interview/adventurebook/repository/BookRepository.java
@@ -1,11 +1,11 @@
 package com.jilds.interview.adventurebook.repository;
 
-import com.jilds.interview.adventurebook.model.entity.BookEntity;
+import com.jilds.interview.adventurebook.domain.entity.BookEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface BookRepository extends JpaRepository<BookEntity, Long>, JpaSpecificationExecutor<BookEntity> {
+public interface BookRepository extends JpaRepository<BookEntity, Integer>, JpaSpecificationExecutor<BookEntity> {
 }
